+++ NEWS.md
@@ -5,6 +5,7 @@
 - New: argument `with` in `exec` to pass processing options that should be preferred over direct naming.
 - New: function `set_exec_options()` to assign global processing options and override arguments potentially hard coded in `exec()`
 - New: stage `load_raster` to read a raster instead of producing it on the fly from the point cloud.
+- New: stage `add_rgb` to modify the point data format
 - Doc: new article on the website about parallelism for an illustrated version of `?multithreading`
 - Doc: improve documentation about processing options in `?exec` and `?set_exec_options`